Run dos2unix and chmod 644
For some or other reason, there were still some carriage returns, and as I'm
now editing in Vim, this is not only unnecessary, but very annoying.

In the past I copied my files to archive from my development board, and I
noticed now all files are still chmod 777 ... This is unneeded, and without
caution to change this on your own board, a security risk.