Delete Add Game Simple, and rename other methods
The Add Game (Simple) method has no function at all, as the tar's them
selves are supported, and the files needed to do a Add Game (Simple) are
exactly those files. The only reasons to continue to support this old
feature are problems with the tar system, but these should be fixed in
future by adding support to use the executable from the server, and by
adding support for other tar-libraries, if possible.