Convert ACP delete games and use short db methods
The deletion of a game has been converted. I noted that I did delete the
tournaments, but I didn't delete the players from the database. This has been
corrected, although I'm a bit afraid how other DBMS systems will react. We'll
see but it should be OK, as this is, I thought, according to the SQL standard.

All other queries in the parts I already converted to the new tables and
standards, I've now altered to use the db-methods of MyBB to the limit.