Make Archive_Tar independent from PEAR
Normally, Archive_Tar should function now properly without the need of the PEAR
library. This is needed as not all servers with php necessarily have PEAR.

The error and warning handling should improve. Now it's simply a die, but this
should change in future. Certainly the warnings should be passed, so they can be
shown to the user. The errors, if possible to stop execution of the
Archive_Tar class, but without destroying it, as we would lose the message then,
we should do the same. I'm not sure, but, we can use the error handling of PHP,
not?