Fix bug updating the tournaments cache
This commit adds some new comments, including TODO's.

When editing a game, the tournament cache should be updated when the active
parameter of the game changed, but the code checked the score type instead...

When editing a game, when there were errors, all changes were lost. Now, when
there were errors, the user input (including the errors) is now filled in the
form.